+++ UI/src/components/features/RipaFormContainer.vue
@@ -1,9 +1,8 @@
 <template>
-  <ripa-page-container :admin="admin">
-    <vue-confirm-dialog></vue-confirm-dialog>
-
+  <div class="ripa-home-container">
     <ripa-form-template
       v-model="stop"
+      :admin-editing="isAdminEditing"
       :beats="mappedFormBeats"
       :county-cities="mappedFormCountyCities"
       :display-beat-input="displayBeatInput"
@@ -39,6 +38,7 @@
       :on-open-template="handleOpenTemplate"
       :on-step-index-change="handleStepIndexChange"
       :on-submit="handleSubmit"
+      :on-update-user="handleUpdateUser"
       @input="handleInput"
     ></ripa-form-template>
 
@@ -92,72 +92,82 @@
       :statute="statute"
       :on-close="handleCloseDialog"
     ></ripa-statute-dialog>
-  </ripa-page-container>
+
+    <ripa-user-dialog
+      :is-invalid-user="isOnlineAndAuthenticated && invalidUser"
+      :user="getMappedUser"
+      :show-dialog="showUserDialog"
+      :on-close="handleClose"
+      :on-save="handleSaveUser"
+    ></ripa-user-dialog>
+  </div>
 </template>
 
 <script>
 import RipaAddFavoriteDialog from '@/components/molecules/RipaAddFavoriteDialog'
+import RipaApiStopJobMixin from '@/components/mixins/RipaApiStopJobMixin'
 import RipaFavoritesDialog from '@/components/molecules/RipaFavoritesDialog'
 import RipaFormTemplate from '@/components/templates/RipaFormTemplate'
-import RipaPageContainer from './RipaPageContainer'
 import RipaStatuteDialog from '@/components/molecules/RipaStatuteDialog'
-import RipaStopMixin from '@/components/mixins/RipaStopMixin'
-import {
-  formBeats,
-  formCountyCities,
-  formNonCountyCities,
-  formSchools,
-  formStatutes,
-} from '../data/mappings'
+import RipaFormContainerMixin from '@/components/mixins/RipaFormContainerMixin'
+import RipaUserDialog from '@/components/molecules/RipaUserDialog'
+import { mapGetters, mapActions } from 'vuex'
 
 export default {
   name: 'ripa-home-container',
 
-  mixins: [RipaStopMixin],
+  mixins: [RipaFormContainerMixin, RipaApiStopJobMixin],
 
   components: {
     RipaAddFavoriteDialog,
     RipaFavoritesDialog,
     RipaFormTemplate,
-    RipaPageContainer,
     RipaStatuteDialog,
+    RipaUserDialog,
   },
 
   data() {
     return {
-      displayBeatInput: true,
-      displayDebugger: true,
-      mappedUser: {
-        agency: 'Insight',
-        startDate: '2010-05-18',
-        yearsExperience: 11,
-        assignment: 1,
-        otherType: null,
-        officerId: '2021050812345',
-        officerName: 'Steve Pietrek',
-      },
-      isAuthenticated: false,
-      isOnlineAndAuthenticated: false,
-      isOnline: true,
-      mappedFormBeats: [],
-      mappedFormCountyCities: [],
-      mappedFormNonCountyCities: [],
-      mappedFormSchools: [],
-      mappedFormStatutes: [],
+      showUserDialog: false,
     }
   },
 
+  computed: {
+    ...mapGetters([
+      'invalidUser',
+      'isOnlineAndAuthenticated',
+      'mappedFormBeats',
+      'mappedFormCountyCities',
+      'mappedFormNonCountyCities',
+      'mappedFormSchools',
+      'mappedFormStatutes',
+      'mappedGpsLocationAddress',
+      'mappedUser',
+      'isAuthenticated',
+      'displayBeatInput',
+      'displayDebugger',
+    ]),
+
+    getMappedUser() {
+      return {
+        agency: this.mappedUser.agency,
+        assignment: this.mappedUser.assignment,
+        otherType: this.mappedUser.otherType,
+        startDate: this.mappedUser.startDate,
+        yearsExperience: this.mappedUser.yearsExperience,
+      }
+    },
+  },
+
   methods: {
-    getFormData() {
-      this.loading = true
-      setTimeout(() => {
-        this.mappedFormSchools = formSchools()
-        this.mappedFormBeats = formBeats()
-        this.mappedFormCountyCities = formCountyCities()
-        this.mappedFormNonCountyCities = formNonCountyCities()
-        this.mappedFormStatutes = formStatutes()
-        this.loading = false
-      }, 500)
+    ...mapActions(['checkTextForPii', 'checkGpsLocation', 'editOfficerUser']),
+
+    handleClose() {
+      this.showUserDialog = false
+    },
+
+    handleSaveUser(user) {
+      this.editOfficerUser(user)
     },
 
     handleSubmit(apiStop) {
@@ -165,12 +175,20 @@ export default {
       this.setLastLocation(this.stop)
     },
 
-    validateLocationForPii(textValue) {
-      if (this.isOnlineAndAuthenticated && textValue && textValue !== '') {
-        const trimmedTextValue = textValue
+    handleUpdateUser() {
+      this.showUserDialog = true
+    },
+
+    async validateLocationForPii(textValue) {
+      const trimmedTextValue = textValue ? textValue.trim() : ''
+      if (
+        this.isOnlineAndAuthenticated &&
+        !this.invalidUser &&
+        trimmedTextValue.length > 0
+      ) {
         this.loadingPiiStep1 = true
         let isFound = false
-        isFound = trimmedTextValue.includes('John Doe')
+        isFound = await this.checkTextForPii(trimmedTextValue)
         this.stop = Object.assign({}, this.stop)
         if (this.stop.location) {
           this.stop.location.piiFound = isFound
@@ -180,11 +198,16 @@ export default {
       }
     },
 
-    validateReasonForStopForPii(textValue) {
-      if (this.isOnlineAndAuthenticated && textValue && textValue !== '') {
+    async validateReasonForStopForPii(textValue) {
+      const trimmedTextValue = textValue ? textValue.trim() : ''
+      if (
+        this.isOnlineAndAuthenticated &&
+        !this.invalidUser &&
+        trimmedTextValue.length > 0
+      ) {
         this.loadingPiiStep3 = true
         let isFound = false
-        isFound = textValue.includes('John Doe')
+        isFound = await this.checkTextForPii(trimmedTextValue)
         this.stop = Object.assign({}, this.stop)
         if (this.stop.stopReason) {
           this.stop.stopReason.reasonForStopPiiFound = isFound
@@ -194,11 +217,16 @@ export default {
       }
     },
 
-    validateBasisForSearchForPii(textValue) {
-      if (this.isOnlineAndAuthenticated && textValue && textValue !== '') {
+    async validateBasisForSearchForPii(textValue) {
+      const trimmedTextValue = textValue ? textValue.trim() : ''
+      if (
+        this.isOnlineAndAuthenticated &&
+        !this.invalidUser &&
+        trimmedTextValue.length > 0
+      ) {
         this.loadingPiiStep4 = true
         let isFound = false
-        isFound = textValue.includes('John Doe')
+        isFound = await this.checkTextForPii(trimmedTextValue)
         this.stop = Object.assign({}, this.stop)
         if (this.stop.actionsTaken) {
           this.stop.actionsTaken.basisForSearchPiiFound = isFound
@@ -209,15 +237,33 @@ export default {
     },
   },
 
-  created() {
-    this.getFormData()
+  watch: {
+    mappedGpsLocationAddress(newVal) {
+      this.lastLocation = newVal
+    },
   },
 
-  props: {
-    admin: {
-      type: Boolean,
-      default: false,
-    },
+  mounted() {
+    const localFormEditing = localStorage.getItem('ripa_form_editing')
+    const localStop = localStorage.getItem('ripa_form_stop')
+    const localFullStop = localStorage.getItem('ripa_form_full_stop')
+    const stepIndex = localStorage.getItem('ripa_form_step_index') || 1
+
+    if (localFormEditing) {
+      const parsedStop = JSON.parse(localStop)
+      const parsedFullStop = JSON.parse(localFullStop)
+
+      this.stop = parsedStop
+      this.fullStop = parsedFullStop
+
+      if (Object.keys(this.fullStop).length > 0) {
+        this.formStepIndex = Number(stepIndex)
+        localStorage.setItem('ripa_form_cached', '1')
+      } else {
+        localStorage.removeItem('ripa_form_admin_editing')
+        localStorage.removeItem('ripa_form_editing')
+      }
+    }
   },
 }
 </script>
